change to js method
